small tweaks
Made the example problem more tractable.
Right now we still have an issue that, with the default solver, the
planner get lost and never produces a solution.

That makes sense as our computation of the distance can only be
computed when the problem is well instantiated but we need to refine
this to do better.

One trick would be to not allow to have “at” waypoints created which
are not fully instantiated. Still the planner can go crazy and make
multiple intermediate waypoints until he realize there’s no way to do
the full coverage.

… I need to figure something out but maybe just ensuring a bounding box
should be enough (?)