Added ability to add new goals
europtus can now integrate new goals the same way it integrates facts.

also added a method that display new plan when planning completes and a
function that allow to test if en europa token is a fact or merged with
a fact.

Next thing si to implement the basics for dispatching goals as we
advance in the execution either based on timing or facts that removes
guards on the execution of an action